\name{EventWinRatios-internal}
\alias{EventWinRatios-internal}
\alias{fn_ds}
\alias{fun_gen}
\alias{fn_gephsim}
\alias{fn_glinsim}
\alias{fn_gmhat}
\alias{fn_wr}
\title{Internal functions for the EventWinRatios package}
\description{
Internal functions for the EventWinRatios package.
}
\value{
No return value, called for side effects
}
